fix: PaddleOCR settings won't save

+++ yomininja-e/electron-src/@core/infra/ocr/ocr_services/paddle_ocr_service/paddle_ocr_service.ts
@@ -357,14 +357,14 @@ export class PaddleOcrService {
         if ( !dirExists )
             fs.mkdirSync( this.settingsPresetsRoot, { recursive: true } );
     
-        // if ( !fileExists ) {
+        if ( !fileExists ) {
             const baseFilePath = join( this.binRoot, '/presets/default.json' );
             const dest = join( this.settingsPresetsRoot, 'default.json' );
             fs.copyFileSync(
                 baseFilePath,
                 dest
             );
-        // }
+        }
     }
 
     handleLanguageTags( tag: string ) {
